feat(cli): added validate to maker cli
